MIRA_MVP0/backend/requirements.txt
# MIRA Backend - Hybrid Lambda + App Runner Deployment
# ============================================================================
# Core FastAPI & ASGI
# ============================================================================
fastapi==0.119.0
uvicorn==0.37.0  # Local dev only (not deployed to Lambda)
mangum==0.17.0  # Lambda WSGI/ASGI adapter
python-multipart==0.0.20
python-dotenv==1.1.1
starlette==0.48.0

# ============================================================================
# AWS SDK (DynamoDB, S3, etc.)
# ============================================================================
boto3>=1.35.0  # AWS SDK (excluded from Lambda, uses built-in version)
aioboto3==13.3.0  # Async AWS operations

# ============================================================================
# Database & Storage
# ============================================================================
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
psycopg2-binary==2.9.10  # PostgreSQL driver
# pgvector - NOT NEEDED (Supabase has it server-side, use RPC functions instead)
postgrest==2.24.0
storage3==2.24.0
gotrue==2.12.4
realtime==2.24.0

# ============================================================================
# Authentication & Security
# ============================================================================
PyJWT==2.10.1
cryptography==46.0.3
# passlib - NOT NEEDED (using Supabase auth instead)
oauthlib==3.3.1
requests-oauthlib==2.0.0

# ============================================================================
# AI & ML
# ============================================================================
openai==2.7.1  # GPT-4, embeddings

# ============================================================================
# Voice & Audio (ElevenLabs)
# ============================================================================
elevenlabs==2.18.0
websockets==13.1  # For ElevenLabs WebSocket STT/TTS
# pydub - NOT NEEDED in Lambda (requires ffmpeg, only needed for local dev)

# ============================================================================
# Google APIs (Gmail, Calendar)
# ============================================================================
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-api-python-client==2.187.0
google-api-core==2.28.1
googleapis-common-protos==1.72.0

# ============================================================================
# Microsoft APIs (Outlook, Calendar) - Optional
# ============================================================================
# msal==1.31.1
# msgraph-sdk==1.14.0

# ============================================================================
# HTTP Clients
# ============================================================================
httpx==0.28.1
requests==2.32.5
aiohttp==3.11.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1  # Local dev only

# ============================================================================
# Data Processing
# ============================================================================
pydantic==2.12.4
pydantic_core==2.41.5
beautifulsoup4==4.14.2
soupsieve==2.8

# ============================================================================
# Date & Time
# ============================================================================
python-dateutil==2.9.0.post0
pytz==2025.2
icalendar==6.3.2
tzdata==2025.2

# ============================================================================
# Payments
# ============================================================================
stripe==6.7.0

# ============================================================================
# Utilities
# ============================================================================
cachetools==6.2.1  # Caching utilities
click==8.3.0
PyYAML==6.0.3

# ============================================================================
# Dependencies (auto-installed with above packages)
# ============================================================================
annotated-types==0.7.0
anyio==4.11.0
awslambdaric==2.0.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.11
jiter==0.11.1
multidict==6.7.0
packaging==25.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.2.5
rsa==4.9.1
simplejson==3.17.2
six==1.17.0
sniffio==1.3.1
StrEnum==0.4.15
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
watchfiles==1.1.1
webencodings==0.5.1
yarl==1.22.0

